Make flat-map-changes consumable from itself.
There was a case where flatMapChanges can be used to write back to
itself.

When this happened in the same stack, writes could occur on a stream
that was already destroyed.

Make data pushes to the output stream async.

This ensures we clear the callstack and do not write to a destroyed
data$.

  - rewrite stream to use pull.
  - make Async.

Signed-off-by: Joe Grund <joe.grund@intel.com>